server/python/requirements.txt
# Web Scraping & Automation (SeleniumBase manages selenium, chromedriver, etc.)
seleniumbase>=4.32.0
playwright>=1.51.0
playwright-stealth>=1.0.6
beautifulsoup4>=4.12.0
pyvirtualdisplay>=3.0  # Virtual display for Linux servers (xvfb)

# Database
psycopg2-binary>=2.9.0
psycopg[binary]>=3.2.0

# Web Framework
flask
flask-cors

# Task Queue & Workers
celery>=5.4.0

# Data Processing
pandas>=2.2.0
numpy>=2.1.0

# HTTP Requests
requests>=2.31.0

# AI/ML
ollama>=0.4.0

# PDF Processing
PyMuPDF>=1.24.0
PyMuPDFb>=1.24.0

# Configuration
python-dotenv>=1.0.0

# Parsing & HTML
lxml>=5.0.0

# Reddit API (uses public .json endpoints, no library needed)